backend/requirements.txt
fastapi==0.104.1
uvicorn==0.38.0
openai==2.8.1
openai-agents==0.6.1
qdrant-client==1.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
alembic==1.17.2
sqlalchemy==2.0.23
fastembed==0.2.2
